added checking of existing blog in updating andd deletion post

+++ 5_sprint_Mongoose_SQL_Nest/src/features/posts/application/use-cases/delete-post.use-case.ts
@@ -7,17 +7,27 @@ import { BlogsQueryRepository } from '../../../blogs/infrastructure/SQL/query.re
 import { modifyPostIntoInitialViewModel } from '../../../../infrastructure/utils/functions/features/posts.functions.helpers';
 import { PostsRepository } from '../../infrastructure/SQL/repository/posts.repository';
 import { PostsOrmRepository } from '../../infrastructure/typeORM/repository/posts-orm.repository';
+import { BlogsOrmQueryRepository } from '../../../blogs/infrastructure/typeORM/query.repository/blogs-orm.query.repository';
 
 export class DeletePostCommand {
-  constructor(public postId: string) {}
+  constructor(public postId: string, public blogId: string) {}
 }
 
 @CommandHandler(DeletePostCommand)
 export class DeletePostUseCase implements ICommandHandler<DeletePostCommand> {
-  constructor(protected postsOrmRepository: PostsOrmRepository) {}
+  constructor(
+    protected postsOrmRepository: PostsOrmRepository,
+    protected blogsOrmQueryRepository: BlogsOrmQueryRepository,
+  ) {}
 
   async execute(command: DeletePostCommand): Promise<boolean> {
-    const { postId } = command;
+    const { postId, blogId } = command;
+
+    const doesBlogExist = await this.blogsOrmQueryRepository.doesBlogExist(
+      blogId,
+    );
+    if (!doesBlogExist) return false;
+
     return this.postsOrmRepository.deleteSinglePost(postId);
   }
 }
